api/requirements.txt
# HTTP / API
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
pydantic[email]>=2.0.0
pydantic-settings>=2.0.0

# Database
SQLAlchemy>=2.0.0
pymysql>=1.1.0

# Auth (bcrypt only — avoids passlib vs bcrypt>=4.1 __about__ breakage)
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0

# Existing integrations
nodriver>=0.48.0
python-dotenv>=1.0.0
httpx>=0.27.0
Pillow>=10.0.0
